General Interface for Partial Least Squares (PLS)

pls() is a way to generate a specification of a model before fitting and allows the model to be created using R. The main arguments for the model are:

num_terms: The number of predictors that are allowed to affect each PLS loading.

num_comp: The number of PLS components to retain.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

pls(mode = "unknown", num_terms = NULL, num_comp = NULL)

# S3 method for pls
update(
  object,
  parameters = NULL,
  num_terms = NULL,
  num_comp = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

num_terms
The number of predictors that are allowed to affect each PLS loading. The default allows each predictor to have non-zero loadings for each PLS component.

num_comp
The number of PLS components to retain.

object
A PLS model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update().

Details

The model can be created using the fit() function using the following engines:

R: "mixOmics" (the default)

Engine Details

Engines may have pre-set default arguments when executing the model fit call. The possible model calls are shown in the Examples section below.

Examples

pls(num_comp = 2, num_terms = 10) %>%
  set_engine("mixOmics") %>%
  set_mode("regression") %>%
  translate()

#> PLS Model Specification (regression)
#> 
#> Main Arguments:
#>   num_terms = 10
#>   num_comp = 2
#> 
#> Computational engine: mixOmics 
#> 
#> Model fit template:
#> plsmod::pls_fit(x = missing_arg(), y = missing_arg(), keepX = 10, 
#>     ncomp = 2)

pls(num_comp = 2, num_terms = 10) %>%
  set_engine("mixOmics") %>%
  set_mode("classification") %>%
  translate()

#> PLS Model Specification (classification)
#> 
#> Main Arguments:
#>   num_terms = 10
#>   num_comp = 2
#> 
#> Computational engine: mixOmics 
#> 
#> Model fit template:
#> plsmod::pls_fit(x = missing_arg(), y = missing_arg(), keepX = 10, 
#>     ncomp = 2)

pls(num_comp = 6) %>%
  set_engine("mixOmics") %>%
  set_mode("regression") %>%
  translate()

#> PLS Model Specification (regression)
#> 
#> Main Arguments:
#>   num_comp = 6
#> 
#> Computational engine: mixOmics 
#> 
#> Model fit template:
#> plsmod::pls_fit(x = missing_arg(), y = missing_arg(), ncomp = 6)

pls() %>%
  set_engine("mixOmics") %>%
  set_mode("classification") %>%
  translate()

#> PLS Model Specification (classification)
#> 
#> Computational engine: mixOmics 
#> 
#> Model fit template:
#> plsmod::pls_fit(x = missing_arg(), y = missing_arg())

model <- pls(num_terms = 10)
model

#> PLS Model Specification (unknown)
#> 
#> Main Arguments:
#>   num_terms = 10
#> 

update(model, num_terms = 1)

#> PLS Model Specification (unknown)
#> 
#> Main Arguments:
#>   num_terms = 1
#> 

update(model, num_terms = 1, fresh = TRUE)

#> PLS Model Specification (unknown)
#> 
#> Main Arguments:
#>   num_terms = 1
#>