Wrapper for mixOmics pls functions

Based on arguments, this wrapper routes the data and arguments to the four pls functions that are sparse/dense or regression/classification.

Usage

pls_fit(x, y, ncomp = NULL, predictor_prop = 1, ...)

Arguments

x

A data frame or matrix of predictors.

y

For classification, a factor. For regression, a matrix, vector, or data frame.

ncomp

The number of PLS components. If left NULL, the maximum possible is used.

predictor_prop

The maximum proportion of original predictors that can have non-zero coefficients for each PLS component (via regularization). This value is used for all PLS components for X.

Value

A model object generated by mixOmics::pls(), mixOmics::plsda(), mixOmics::spls(), or mixOmics::splsda().